fastapi/requirements.txt
fastapi
httpx
aiofiles
pydantic
motor
pymongo
uvicorn
django-environ
redis